refactored a bit around ts safety and moved all source code in the right folders.
