+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.ts", 
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "tsc --module commonjs && node dist/index.js"
+    "start": "tsc --module commonjs && node dist/src/index.js"
   },
   "author": "",
   "license": "ISC",
