+++ src/helpers/openai.ts
